+++ tests/test_aliens.py
@@ -1,6 +1,7 @@
 import pytest
 
-from app.backend.models.alien import Alien, Team
+from app.backend.models.alien import Alien
+from app.backend.models.team import Team
 
 
 @pytest.fixture
